LPS-55928 move tokenizer to analysis package
