+++ portal-service/src/com/liferay/portal/kernel/search/suggest/BaseQuerySuggester.java
@@ -18,7 +18,7 @@ import com.liferay.portal.kernel.search.DocumentImpl;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.search.SearchException;
-import com.liferay.portal.kernel.search.TokenizerUtil;
+import com.liferay.portal.kernel.search.analysis.TokenizerUtil;
 
 import java.util.List;
 import java.util.Map;
